requirements.txt
python-dotenv==0.9.1
python-telegram-bot==12.2
Telethon==1.10.9
